redux/state: move app actions, reducer under store/
