+++ webapp/store/app/index.js
@@ -1,1 +1,4 @@
+import * as AppActions from './actions'
+
 export { useAssessmentType, useCountries, useCountriesPanEuropean, useRegions } from './hooks'
+export { AppActions }
