+++ webapp/store/user/reducer.js
@@ -3,7 +3,7 @@ import { exportReducer } from '@webapp/utils/reduxUtils'
 import * as UserState from '@webapp/store/user/state'
 import * as UserActions from '@webapp/store/user/actions'
 
-import { appI18nUpdate, appInitDone } from '@webapp/app/actions'
+import { appI18nUpdate, appInitDone } from '@webapp/store/app/actions'
 
 const actionHandlers = {
   [appInitDone]: (state, { userInfo }) => UserState.assocUserInfo(userInfo)(state),
